requirements.txt
asgiref==3.3.1
Django==3.1.4
django-ckeditor==6.0.0
django-filter==2.4.0
django-js-asset==1.2.2
django-wysiwyg==0.8.0
djangorestframework==3.12.2
Pillow==8.0.1
pytz==2020.4
selenium==3.141.0
sqlparse==0.4.1
urllib3==1.26.2
xxx==0.0.1
